+++ src/main/java/com/github/cc3002/finalreality/model/weapon/AbstractWeapon.java
@@ -8,7 +8,7 @@ import java.util.Objects;
  * @author Ignacio Slater Muñoz.
  * @author Camila Labarca
  */
-public abstract class AbstractWeapon {
+public abstract class AbstractWeapon implements IWeapon {
 
   private final String name;
   private final int damage;