+++ src/test/java/com/github/cc3002/finalreality/model/character/AbstractCharacterTest.java
@@ -1,14 +1,10 @@
 package com.github.cc3002.finalreality.model.character;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotEquals;
-
-import java.util.ArrayList;
-import java.util.List;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.LinkedBlockingQueue;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
 
 /**
  * Abstract class containing the common tests for all the types of characters.
@@ -21,14 +17,13 @@ import org.junit.jupiter.api.Test;
 public abstract class AbstractCharacterTest {
 
   protected BlockingQueue<ICharacter> turns;
-  protected List<ICharacter> testCharacters;
 
   /**
    * Checks that the character waits the appropriate amount of time for it's turn.
    */
-  @Test
-  void waitTurnTest() {
-    testCharacters.get(0).waitTurn();
+  protected void checkWaitTurn(ICharacter character) {
+    Assertions.assertTrue(turns.isEmpty());
+    character.waitTurn();
     try {
       // Thread.sleep is not accurate so this values may be changed to adjust the
       // acceptable error margin.
@@ -37,27 +32,32 @@ public abstract class AbstractCharacterTest {
       Assertions.assertEquals(0, turns.size());
       Thread.sleep(200);
       Assertions.assertEquals(1, turns.size());
-      Assertions.assertEquals(testCharacters.get(0), turns.peek());
+      Assertions.assertEquals(character, turns.peek());
     } catch (InterruptedException e) {
       e.printStackTrace();
     }
   }
 
   protected void checkConstruction(final ICharacter expectedCharacter,
-      final ICharacter testEqualCharacter,
-      final ICharacter sameClassDifferentCharacter,
-      final ICharacter differentClassCharacter) {
+                                   final ICharacter testEqualCharacter,
+                                   final ICharacter sameClassDifferentCharacter,
+                                   final ICharacter differentClassCharacter) {
     assertEquals(expectedCharacter, testEqualCharacter);
     assertNotEquals(sameClassDifferentCharacter, testEqualCharacter);
     assertNotEquals(testEqualCharacter, differentClassCharacter);
     assertEquals(expectedCharacter.hashCode(), testEqualCharacter.hashCode());
-    assertEquals(expectedCharacter.getPoints(), testEqualCharacter.getPoints());
-    assertEquals(expectedCharacter.getDefense(), testEqualCharacter.getDefense());
   }
 
 
   protected void basicSetUp() {
     turns = new LinkedBlockingQueue<>();
-    testCharacters = new ArrayList<>();
   }
+
+  protected void checkAttack(ICharacter attacker, ICharacter attacked){
+    assertTrue(attacker.isAlive());
+    assertTrue(attacked.isAlive());
+    attacker.attack(attacked);
+
+  }
+
 }