+++ ChangeLog.md
@@ -4,6 +4,12 @@ ChangeLog
 Version 2.0
 -----------
 
+- (RC.4) Game controller tests
+- (B.13) Game controller
+- (RC.3) Tests for attacks and equipment
+- (B.12) Implemented weapon's equipment with double dispatch
+- (B.11) Created IWeapon interface
+- (B.10) Added attacks to characters with double dispatch
 - (RC.2) Modified weapons tests
 - (B.9) Changed weapon to abstract class and created classes for each weapon
 - (RC.1) Modified tests